Adding course DONE

+++ server/controllers/Section.js
@@ -1,6 +1,6 @@
 const Section = require("../models/Section")
 const Course = require("../models/Course")
-
+const SubSection = require("../models/SubSection")
 exports.createSection = async(req,res)=>{
     try {
         // data fetch 
@@ -49,13 +49,21 @@ exports.createSection = async(req,res)=>{
 exports.updateSection  = async(req,res)=>{
     try {
         // data input 
-        const {sectionName, sectionId} = req.body;
+        const {sectionName, sectionId, courseId} = req.body;
         // update data 
-        const section  = await Section.findByIdAndUpdate(sectionId, {sectionName},  {new:true}
-            )
-        // return respinse
+        const section  = await Section.findByIdAndUpdate(sectionId, {sectionName},  {new:true})
+
+        const course = await Course.findById(courseId).populate({
+            path: "courseContent",
+            populate:{
+                path:"subSection"
+            }
+        })
+        .exec();
+        // return response
         return res.status(200).json({
             success:true,
+            data: course,
             message:section,
         })
     } catch (error) {
@@ -68,25 +76,71 @@ exports.updateSection  = async(req,res)=>{
 }
  
 
-exports.deleteSection = async(req,res)=>{
-    try {
-        // get id -- assuming we are sending id in params 
-        // check with req.params
-        const {sectionId} = req.body
+// exports.deleteSection = async(req,res)=>{
+//     try {
+//         // get id -- assuming we are sending id in params 
+//         // check with req.params
+//         const {sectionId} = req.body
 
-        // delete section 
-        await Section.findByIdAndDelete(sectionId) 
+//         // delete section 
+//         await Section.findByIdAndDelete(sectionId) 
 
-        // return response
-        return res.status(200).json({
-            success:true,
-            message:"Section deleted successfully"
-        })
-    } catch (error) {
-        return res.status(500).json({
-            success:false,
-            message:"Unable to update section, please try again",
-            error:error.message
-        })
-    }
-}
+//         // return response
+//         return res.status(200).json({
+//             success:true,
+//             message:"Section deleted successfully"
+//         })
+//     } catch (error) {
+//         return res.status(500).json({
+//             success:false,
+//             message:"Unable to update section, please try again",
+//             error:error.message
+//         })
+//     }
+// }
+
+exports.deleteSection = async (req, res) => {
+	try {
+
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
+
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
+};
